Allow StacheDefineElement to be used with import/export transpilation
This allows StacheDefineElement to be used with import/export syntax in
loaders/bundlers that transpile to ES5, like steal.

ES2015 class constructors cannot be called with `.call`, but that's how
Babel 6 transpiles classes to constructor functions. The trick is to
insert a constructor function so that the ES2015 classes are not called
with `.call`.

Ideally this code could go somewhere like can-reflect where it could be
used in other projects.