Ensure that both addEventListeners are called
This ensure that events registered with `addEventListener` are
registered in both the CanJS event system and also in the DOM event
system. This means that both DOM events and property events are in both
systems, and can be listened to the same way.

The downside is that you can't have a `click` property and a `click`
event and listen to those as different things, because now they will not
be.